+++ packages/config-manager/src/index.ts
@@ -1,5 +1,14 @@
 export * from "./types";
 export { initializeConfig, getConfig, validateConfig } from "./manager";
+/**
+ * @deprecated use the {@link nameOfScript} and {@link findConfigByScript} function instead
+ */
 export * as helpers from "./helpers";
+export { nameOfScript, findConfigByScript } from "./helpers";
 export { predefined, createConfig } from "./predefined";
 export { generateGenesisScriptConfigs } from "./genesis";
+export {
+  refreshScriptConfigs,
+  createRpcResolver,
+  createLatestTypeIdResolver,
+} from "./refresh";