+++ packages/lumos/src/common-scripts/omnilock.ts
@@ -1,0 +1,9 @@
+export {
+  bitcoin,
+  OmnilockWitnessLock,
+  createOmnilockScript,
+  setupInputCell,
+  CellCollector,
+  prepareSigningEntries,
+} from "@ckb-lumos/common-scripts/lib/omnilock";
+export type { OmnilockInfo } from "@ckb-lumos/common-scripts/lib/omnilock";