+++ packages/base/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ckb-lumos/base
 
+## 0.22.0-next.3
+
+## 0.22.0-next.2
+
+### Minor Changes
+
+- [#601](https://github.com/ckb-js/lumos/pull/601) [`7e89b4a`](https://github.com/ckb-js/lumos/commit/7e89b4a1b382e9c6fb4e1305bdaf1521bc0cde6e) Thanks [@homura](https://github.com/homura)! - BREAKING CHANGE: unpacked number of `blockchain.*.unpack` should be `Hexadecimal` instead of `number` or`BI`
+
 ## 0.22.0-next.1
 
 ## 0.22.0-next.0