+++ packages/config-manager/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ckb-lumos/config-manager
 
+## 0.22.0-next.3
+
+### Minor Changes
+
+- [#610](https://github.com/ckb-js/lumos/pull/610) [`0353ba3`](https://github.com/ckb-js/lumos/commit/0353ba357454255703ba86f98b44333c6c966492) Thanks [@homura](https://github.com/homura)! - feat: `refreshScriptConfigs` to refresh deployment config
+
+- [#606](https://github.com/ckb-js/lumos/pull/606) [`349be74`](https://github.com/ckb-js/lumos/commit/349be74dd828c68036373c8eae6dd912c711551f) Thanks [@homura](https://github.com/homura)! - feat: **deprecated** the `helpers` in `config-manager` to avoid confused with `@ckb-lumos/helpers`
+
+### Patch Changes
+
+- [#614](https://github.com/ckb-js/lumos/pull/614) [`5ee0349`](https://github.com/ckb-js/lumos/commit/5ee0349c5d0d5294acafe4788fd225609bfc7235) Thanks [@homura](https://github.com/homura)! - feat: update omnilock script config
+
+## 0.22.0-next.2
+
+### Minor Changes
+
+- [#605](https://github.com/ckb-js/lumos/pull/605) [`330ec9a`](https://github.com/ckb-js/lumos/commit/330ec9a79dc024e7860aa4785fe9c3bcfffc0767) Thanks [@homura](https://github.com/homura)! - feat: update testnet omnilock config
+
 ## 0.22.0-next.1
 
 ## 0.22.0-next.0